2024.09.22-23 문현동 작업 문현동 작업내역: 1. 각종 페이지 + 버튼 등 요소들 디자인. 2. 메세지 알림에 받은 메세지의 내용 나오게 하기. 문현동 해야할 일: 1. 알림 페이지 -> 알림 모달로 만들기. 2. 전체적인 이식 결과 테스트. 주의사항: 작업 특성상 이원준 개발자의 작업물을 건드릴 수 밖에 없는 작업이어서 몇 군데 변경점이 있고...이에 대해 merge시 충돌 가능성 있음.
